Apply Saga to 3 projects. Test ok.
