--- product-service/src/main/java/com/example/productservice/ProductServiceApplication.java
@@ -1,15 +1,0 @@
-package com.example.productservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@EnableDiscoveryClient
-@SpringBootApplication
-public class ProductServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProductServiceApplication.class, args);
-    }
-
-}
