--- product-service/src/main/java/com/example/productservice/business/command/api/CreateProductCommand.java
@@ -1,17 +1,0 @@
-package com.example.productservice.business.command.api;
-
-import lombok.Data;
-import org.axonframework.modelling.command.TargetAggregateIdentifier;
-
-import java.math.BigDecimal;
-
-@Data
-public class CreateProductCommand {
-
-    @TargetAggregateIdentifier
-    private String productId;
-    private String title;
-    private BigDecimal price;
-    private int quantity;
-
-}
