--- product-service/src/main/java/com/example/productservice/business/config/AxonConfigSecurity.java
@@ -1,19 +1,0 @@
-package com.example.productservice.business.config;
-
-import com.thoughtworks.xstream.XStream;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class AxonConfigSecurity {
-
-    @Bean
-    public XStream xStream() {
-        XStream xStream = new XStream();
-
-        xStream.allowTypesByWildcard(new String[] {
-                "org.msss.cqrs.saga.sharedcommon.**"
-        });
-        return xStream;
-    }
-}
