--- product-service/src/main/java/com/example/productservice/business/event/api/ProductCreateEvent.java
@@ -1,15 +1,0 @@
-package com.example.productservice.business.event.api;
-
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Data
-public class ProductCreateEvent {
-
-    private String productId;
-    private String title;
-    private BigDecimal price;
-    private int quantity;
-
-}
