--- product-service/src/main/java/com/example/productservice/business/event/api/ProductRepo.java
@@ -1,9 +1,0 @@
-package com.example.productservice.business.event.api;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepo extends JpaRepository<ProductEntity, String> {
-
-    ProductEntity findProductEntitiesByProductIdOrTitle(String id, String title);
-
-}
-
