--- product-service/src/main/java/com/example/productservice/business/exception/CommonEventException.java
@@ -1,15 +1,0 @@
-package com.example.productservice.business.exception;
-
-import org.axonframework.eventhandling.EventMessage;
-import org.axonframework.eventhandling.EventMessageHandler;
-import org.axonframework.eventhandling.ListenerInvocationErrorHandler;
-
-import javax.annotation.Nonnull;
-
-public class CommonEventException implements ListenerInvocationErrorHandler {
-
-    @Override
-    public void onError(@Nonnull Exception exception, @Nonnull EventMessage<?> event, @Nonnull EventMessageHandler eventHandler) throws Exception {
-        throw exception;
-    }
-}
