--- product-service/src/main/java/com/example/productservice/business/exception/ErrorMessage.java
@@ -1,16 +1,0 @@
-package com.example.productservice.business.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@AllArgsConstructor
-public class ErrorMessage {
-
-    private LocalDateTime localdatetime;
-    private String message;
-    private String details;
-
-}
