--- product-service/src/main/java/com/example/productservice/business/query/api/FindProductsQuery.java
@@ -1,4 +1,0 @@
-package com.example.productservice.business.query.api;
-
-public class FindProductsQuery {
-}
