--- product-service/src/main/java/com/example/productservice/business/query/api/ProductRestModel.java
@@ -1,15 +1,0 @@
-package com.example.productservice.business.query.api;
-
-
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Data
-public class ProductRestModel {
-
-    private String productId;
-    private String title;
-    private BigDecimal price;
-    private int quantity;
-}
