+++ product-service/src/main/java/org/msss/cqrs/saga/productservice/ProductServiceApplication.java
@@ -1,0 +1,18 @@
+package org.msss.cqrs.saga.productservice;
+
+import org.msss.cqrs.saga.productservice.business.config.AxonConfigSecurity;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.Import;
+
+@EnableDiscoveryClient
+@SpringBootApplication
+@Import({AxonConfigSecurity.class})
+public class ProductServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProductServiceApplication.class, args);
+    }
+
+}
