+++ product-service/src/main/java/org/msss/cqrs/saga/productservice/business/BaseProductController.java
@@ -1,0 +1,21 @@
+package org.msss.cqrs.saga.productservice.business;
+
+import org.axonframework.commandhandling.gateway.CommandGateway;
+import org.axonframework.queryhandling.QueryGateway;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/products")
+public class BaseProductController {
+
+    protected final CommandGateway commandGateway;
+    protected final QueryGateway queryGateway;
+
+    public BaseProductController(CommandGateway commandGateway, QueryGateway queryGateway){
+        this.commandGateway =  commandGateway;
+        this.queryGateway =  queryGateway;
+    }
+
+}
