+++ product-service/src/main/java/org/msss/cqrs/saga/productservice/business/command/lookup/ProductLookUpEntity.java
@@ -1,0 +1,20 @@
+package org.msss.cqrs.saga.productservice.business.command.lookup;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.*;
+
+@AllArgsConstructor
+@Entity
+@Table(name = "product_lookup")
+public class ProductLookUpEntity {
+
+    public ProductLookUpEntity(){};
+
+    @Id
+    private String productId;
+
+    private String title;
+}
