+++ product-service/src/main/java/org/msss/cqrs/saga/productservice/business/exception/CommonException.java
@@ -1,0 +1,46 @@
+package org.msss.cqrs.saga.productservice.business.exception;
+
+import org.axonframework.commandhandling.CommandExecutionException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+
+import java.time.LocalDateTime;
+
+@ControllerAdvice
+public class CommonException {
+
+    @ExceptionHandler(value = {IllegalArgumentException.class})
+    public ResponseEntity<ErrorMessage> handleIllegalStateException(IllegalArgumentException ex, WebRequest request) {
+        ErrorMessage errorMessage = new ErrorMessage(LocalDateTime.now(),
+                ex.getMessage(), request.getDescription(false));
+        return new ResponseEntity<ErrorMessage>(errorMessage, HttpStatus.UNAUTHORIZED);
+    }
+
+    @ExceptionHandler(value =  {Exception.class})
+    public ResponseEntity<ErrorMessage> handleExceptionChung(Exception ex, WebRequest request) {
+        ErrorMessage errorMessage = new ErrorMessage(LocalDateTime.now(),
+                ex.getMessage(), request.getDescription(false));
+        return new ResponseEntity<ErrorMessage>(errorMessage, HttpStatus.UNAUTHORIZED);
+    }
+
+
+    @ExceptionHandler(value =  {ArrayIndexOutOfBoundsException.class})
+    public ResponseEntity<ErrorMessage> testIfEventHandlerIsHere(ArrayIndexOutOfBoundsException ex, WebRequest request) {
+        ErrorMessage errorMessage = new ErrorMessage(LocalDateTime.now(),
+                ex.getMessage(), request.getDescription(false));
+        return new ResponseEntity<ErrorMessage>(errorMessage, HttpStatus.UNAUTHORIZED);
+    }
+
+    @ExceptionHandler(value =  {CommandExecutionException.class})
+    public ResponseEntity<ErrorMessage> testCommandExecutionException(CommandExecutionException ex, WebRequest request) {
+        ErrorMessage errorMessage = new ErrorMessage(LocalDateTime.now(),
+                ex.getMessage(), request.getDescription(false));
+        return new ResponseEntity<ErrorMessage>(errorMessage, HttpStatus.UNAUTHORIZED);
+    }
+
+
+
+}
