+++ product-service/src/main/java/org/msss/cqrs/saga/productservice/business/exception/ErrorMessage.java
@@ -1,0 +1,16 @@
+package org.msss.cqrs.saga.productservice.business.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+public class ErrorMessage {
+
+    private LocalDateTime localdatetime;
+    private String message;
+    private String details;
+
+}
