+++ product-service/src/main/java/org/msss/cqrs/saga/productservice/business/query/api/FindProductsQuery.java
@@ -1,0 +1,4 @@
+package org.msss.cqrs.saga.productservice.business.query.api;
+
+public class FindProductsQuery {
+}
