--- product-service/src/test/java/com/example/productservice/ProductServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.productservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProductServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
