+++ product-service/src/test/java/org/msss/cqrs/saga/productservice/ProductServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package org.msss.cqrs.saga.productservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
